Consolidate physical process controller tests
